+++ src/Widget.php
@@ -103,11 +103,7 @@ abstract class Widget implements NoEncodeStringableInterface, Stringable
      */
     final public function render(): string
     {
-        if (!$this->beforeRun()) {
-            return '';
-        }
-
-        return (string) $this->afterRun((string) $this->run());
+        return (string) $this->run();
     }
 
     /**
@@ -129,56 +125,4 @@ abstract class Widget implements NoEncodeStringableInterface, Stringable
      * when implementing concrete widget.
      */
     abstract protected function run(): string|Stringable;
-
-    /**
-     * This method is invoked right before the widget is executed.
-     *
-     * The return value of the method will determine whether the widget should continue to run.
-     *
-     * When overriding this method, make sure you call the parent implementation like the following:
-     *
-     * ```php
-     * protected function beforeRun(): bool
-     * {
-     *     if (!parent::beforeRun()) {
-     *         return false;
-     *     }
-     *
-     *     // your custom code here
-     *
-     *     return true; // or false to not run the widget
-     * }
-     * ```
-     *
-     * @return bool Whether the widget should continue to be executed.
-     */
-    protected function beforeRun(): bool
-    {
-        return true;
-    }
-
-    /**
-     * This method is invoked right after a widget is executed.
-     *
-     * The return value of the method will be used as the widget return value.
-     *
-     * If you override this method, your code should look like the following:
-     *
-     * ```php
-     * protected function afterRun(string $result): string|\Stringable
-     * {
-     *     $result = parent::afterRun($result);
-     *     // your custom code here
-     *     return $result;
-     * }
-     * ```
-     *
-     * @param string $result The widget return result.
-     *
-     * @return string|Stringable The processed widget result.
-     */
-    protected function afterRun(string $result): string|Stringable
-    {
-        return $result;
-    }
 }